consumer/requirements.txt
google-cloud-pubsub==2.19.0
pytest==7.4.3